Add messageFactory
