+++ components/app/conversation.js
@@ -3,9 +3,8 @@ import $ from '../../script/ui/DOM.js';
 /* eslint-disable */
 import UIIcon from '../ui/icon.js'
 import PeerAvatar from '../ui/peer-avatar.js'
-import AppMessageGroup from '../app/message-group.js'
 import {getDialogTitle, previewMessage} from '../../state/dialogs/helpers.js'
-import {dateDay, formatDate} from '../../script/helpers.js'
+import {dateDay, formatDate, getTimestamp} from '../../script/helpers.js'
 import {setActiveDialog} from '../../state/dialogs/actions.js'
 import {getDialogWithLastMessage$} from '../../state/dialogs/stream-builders.js'
 import {getUserFullName} from '../../state/users/utils.js';
@@ -203,7 +202,7 @@ const getIdFromPeer = R.pipe(
       const chaption = getDialogTitle(dialog);
       const lastMessageDate = R.pathOr(0, ['last_message', 'date'], dialog);
       const updated  = formatDate(dateDay(1000 * lastMessageDate), true);
-      const timestamp = current === updated ? AppMessageGroup.timestamp(lastMessageDate) : updated;
+      const timestamp = current === updated ? getTimestamp(lastMessageDate) : updated;
 
       // pinned, muted
